Fix what looks like a possible crash in input code after at least one player has died. (I haven't seem a crash from myself, but looks likely.)
I missed the fact that p_players is not zeroed after a player is deleted, and instead a second array is used the handle the dangling pointers.

An array like this isn't really normal or safe. Would be better to null the pointers, or use a smart pointer or something like this http://corrodinggames.com/forums/viewtopic.php?f=2&t=12 :p